requirements.txt
os
re
time
pathlib
tqdm>=4.4.0
chardet>=4.0.0
urllib3>=1.26.18
requests>=2.31.0
